Generic instance optimizations
Don't clear generic shader instance buffer on map load. Wait for it to fill up first, then nuke it. Allows one to put most of the stock tagset's generic shaders in the ui.map and have them essentially precompile on game launch so between level loads are not waiting on shader compile.